working delete, needs to be stateful
