+++ resources/js/Map/popUp/showPopUp.jsx
@@ -2,7 +2,7 @@ import mapboxgl from "mapbox-gl";
 import "./pop-up.scss";
 import { getPins } from "../Pins/getPins";
 import { pinOnMap } from "../Pins/addPinOnMap/addPinOnMap";
-import DeletePin from "../Pins/DeletePin";
+import { DeletePin } from "../Pins/DeletePin";
 import React, { useEffect, useState, useCallback } from "react";
 import Modal from "../../components/Modal/Modal";
 import useToggle from "../../components/Modal/use-toggle";
@@ -22,13 +22,9 @@ export default function ShowPopUp({ map }) {
         id: "",
         images: "",
     });
-
-    // const handleEditClick = () => {
-    //     handleToggleModal(<MyFormModalContent details={details} />);
-    // };
-
-    const handleDeleteClick = () => {};
-
+    const handleDeleteClick = () => {
+        DeletePin(details.id);
+    };
     const handleClick = useCallback(
         (e) => {
             const pinProperties = e.features[0].properties;
